+++ packages/nextjs/components/scaffold-eth/index.tsx
@@ -4,3 +4,4 @@ export { default as Address } from "./Address";
 export { default as Balance } from "./Balance";
 export { default as BlockieAvatar } from "./BlockieAvatar";
 export { default as ContractUI } from "./Contract/ContractUI";
+export * from "./Input";
